Add doctrine odm support to be able to use mongodb #70
