+++ src/AerialShip/SamlSPBundle/Bridge/LogoutBase.php
@@ -3,7 +3,7 @@
 namespace AerialShip\SamlSPBundle\Bridge;
 
 use AerialShip\SamlSPBundle\Config\ServiceInfo;
-use AerialShip\SamlSPBundle\State\SSO\SSOState;
+use AerialShip\SamlSPBundle\Model\SSOState;
 use AerialShip\SamlSPBundle\State\SSO\SSOStateStoreInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Http\HttpUtils;
