+++ src/AerialShip/SamlSPBundle/Resources/config/orm.yml
@@ -8,7 +8,7 @@ services:
     aerial_ship_saml_sp.entity_manager:
         class: Doctrine\ORM\EntityManager
         arguments:
-            - %aerial_ship_saml_sp.model_manager_name% # not realy used since factory is set in extension
+            - %aerial_ship_saml_sp.model_manager_name% # used by factory that is set in extension
         public: false
 
     aerial_ship_saml_sp.state.store.sso:
