+++ src/AerialShip/SamlSPBundle/Resources/doc/index.md
@@ -18,6 +18,7 @@ Installation is quick 4 steps:
 4. Configure SamlSpBundle
 5. Configure application's security.yml
 6. Import SamlSpBundle routing
+7. Exchange metadata
 
 
 Step 1: Download SamlSpBundle with composer
@@ -91,16 +92,6 @@ class SSOState extends \AerialShip\SamlSPBundle\Model\SSOState
 }
 ```
 
-After the entity class is created you can create a db migration
-``` bash
-$ php app/console doctrine:migrations:diff
-```
-
-or if you really don't use migrations, just update your database schema by running
-``` bash
-$ php app/console doctrine:schema:update --force
-```
-
 
 Step 4: Configure SamlSpBundle
 ------------------------------
@@ -114,7 +105,23 @@ aerial_ship_saml_sp:
     sso_state_entity_class: AppBundle\Entity\SSOState
 ```
 
-If you are using a manager other then default, you can set it's name in ``model_manager_name`` config
+If you are using a manager other then default one, you can set it's name with ``model_manager_name`` config
+
+``` yaml
+# app/config/config.yml
+aerial_ship_saml_sp:
+    model_manager_name: my_manager
+```
+
+Now you can create a db migration
+``` bash
+$ php app/console doctrine:migrations:diff
+```
+
+or if you really don't use migrations, just update your database schema by running
+``` bash
+$ php app/console doctrine:schema:update --force
+```
 
 
 Step 5: Configure application's security.yml
@@ -189,7 +196,9 @@ aerialship_saml_sp_bundle:
 Step 7: Exchange metadata
 -------------------------
 
-Download your SP metadata by visiting the (configurable) URL `/saml/sp/FederationMetadata.xml` and send the file to the IdP. Save the IdP metadata in your bundle at the configured location (e.g. `@AcmeSamlBundle/Resources/idp-FederationMetadata.xml`).
+Download your SP metadata by visiting the (configurable) URL `/saml/sp/FederationMetadata.xml` and send the file to
+the IDP. Save the IDP metadata in your bundle at the configured location
+(e.g. `@AcmeSamlBundle/Resources/idp-FederationMetadata.xml`).
 
 
 Next Steps
